абстрактный класс и простые элементы
